% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_row_add.R
\name{add_row_ui}
\alias{add_row_ui}
\title{Add row Module UI}
\usage{
add_row_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI with buttons to add rows.
}
\description{
A Shiny module UI for adding rows in the table.
}
