% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_add_table.R
\name{add_table_server}
\alias{add_table_server}
\title{Add Table Module Server}
\usage{
add_table_server(id, json_data, selected_table_name, update_table_choices)
}
\arguments{
\item{id}{Module ID}

\item{json_data}{A reactive value containing the JSON data (list of tables).}

\item{selected_table_name}{A reactive value containing the name of the currently selected table.}

\item{update_table_choices}{A function to update the table choices in the selector dropdown.}
}
\value{
None. The function modifies the `json_data` reactive value in place.
}
\description{
A Shiny module server for adding tables to the JSON data.
}
