% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_add.R
\name{add_table_ui}
\alias{add_table_ui}
\title{Add table Module UI}
\usage{
add_table_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI with buttons to add tables.
}
\description{
A Shiny module UI for adding tables in the JSON data.
}
