% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_mapping_table}
\alias{create_mapping_table}
\title{Handles logic for creating mapping tables according to the mapping type}
\usage{
create_mapping_table(
  full_data,
  table_name,
  table_type = c("Implicit", "Fixed", "Mapping")
)
}
\arguments{
\item{full_data}{A named list representing the full dataset. This is where the new table will be added.}

\item{table_name}{A string specifying the name of the new table to be created.}

\item{table_type}{A string indicating the type of the table to be created.
Must be one of `"Implicit"`, `"Fixed"`, or `"Mapping"`. Defaults to `"Implicit"`.}
}
\value{
The modified `full_data` list with the new table added, or the original `full_data` if `table_type` is `"Implicit"`.
}
\description{
This function adds a new table or update an existing one in the `full_data` list based on the specified table type.
}
\details{
- If `table_type` is `"Fixed"`, the new table is a data frame with one column: `FIXED`.
- If `table_type` is `"Mapping"`, the new table is a data frame with two columns: `SOURCE` and `TARGET`.
- If `table_type` is `"Implicit"`, the function returns the `full_data` unchanged.
}
\examples{
# Example full dataset
full_data <- list()

# Create a Fixed table
full_data <- create_mapping_table(full_data, "FixedTable", "Fixed")

# Create a Mapping table
full_data <- create_mapping_table(full_data, "MappingTable", "Mapping")

# No table is added for "Implicit"
full_data <- create_mapping_table(full_data, "NoTable", "Implicit")

}
