% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_define_mapping_type.R
\name{define_mapping_type_modal}
\alias{define_mapping_type_modal}
\title{Creates a modal for user to define the mapping type}
\usage{
define_mapping_type_modal(
  ns,
  title = "Define Table Type",
  table_name,
  table_types = c("Fixed", "Mapping"),
  selected_type = "Fixed"
)
}
\arguments{
\item{ns}{A namespace function, typically `session$ns`, to ensure unique IDs in the Shiny app.}

\item{title}{The title of the modal dialog. Defaults to "Define Table Type".}

\item{table_name}{The name of the selected table (to be displayed to the user).}

\item{table_types}{A character vector of table type options. Defaults to c("Fixed", "Mapping").}

\item{selected_type}{The default selected table type. Defaults to "Fixed".}
}
\value{
A Shiny modal dialog UI object.
}
\description{
Creates a modal dialog for defining the type of mapping.
}
