% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_define_mapping_type.R
\name{define_mapping_type_server}
\alias{define_mapping_type_server}
\title{Table Definition Module Server}
\usage{
define_mapping_type_server(id, json_data, selected_table_name)
}
\arguments{
\item{id}{Module ID}

\item{json_data}{A reactive value containing the JSON data.}

\item{selected_table_name}{A reactive expression for the selected table name.}
}
\value{
Nothing; modifies `json_data` in place when a table type is defined.
}
\description{
Table Definition Module Server
}
