% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_define_mapping_type.R
\name{define_mapping_type_ui}
\alias{define_mapping_type_ui}
\title{Table Definition Module UI}
\usage{
define_mapping_type_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for the table definition module.
}
\description{
Table Definition Module UI
}
