% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_row_delete.R
\name{delete_row_ui}
\alias{delete_row_ui}
\title{Delete row Module UI}
\usage{
delete_row_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI with buttons to delete rows.
}
\description{
A Shiny module UI for deleting rows in the table.
}
