% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_delete_table.R
\name{delete_table_ui}
\alias{delete_table_ui}
\title{Delete Table Module UI}
\usage{
delete_table_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI with a button to trigger the deletion of a table.
}
\description{
A Shiny module UI for deleting tables in the JSON data.
}
