% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download.R
\name{download_server}
\alias{download_server}
\title{Download Module Server}
\usage{
download_server(id, json_data, file_path)
}
\arguments{
\item{id}{Module ID}

\item{json_data}{A reactive value containing the JSON data.}

\item{file_path}{A reactive value containing the original file path.}
}
\value{
None.
}
\description{
A Shiny module server for handling JSON file download.
}
