% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download.R
\name{download_ui}
\alias{download_ui}
\title{Download Module UI}
\usage{
download_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A `downloadButton` UI element.
}
\description{
A Shiny module UI for downloading the updated JSON file.
}
