% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_edit_table.R
\name{edit_table_ui}
\alias{edit_table_ui}
\title{Table Editor Module UI}
\usage{
edit_table_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A `DTOutput` for displaying the editable table.
}
\description{
A Shiny module UI for displaying and editing the selected table.
}
