% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_loader.R
\name{file_loader_server}
\alias{file_loader_server}
\title{File Loader Module Server}
\usage{
file_loader_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A list containing reactive values `json_data` and `file_path`.
}
\description{
A Shiny module server for handling JSON file uploads and parsing.
}
