% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_loader.R
\name{file_loader_ui}
\alias{file_loader_ui}
\title{File Loader Module UI}
\usage{
file_loader_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A `fileInput` UI element for selecting a JSON file.
}
\description{
A Shiny module UI for uploading JSON files.
}
