% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_add_table.R
\name{generate_add_table_modal}
\alias{generate_add_table_modal}
\title{Generate UI for Add Table Modal}
\usage{
generate_add_table_modal(
  ns,
  title = "Add New Table",
  table_types = c("Fixed", "Mapping"),
  selected_type = "Fixed"
)
}
\arguments{
\item{ns}{A namespace function, typically `session$ns`, to ensure unique IDs in the Shiny app.}

\item{title}{The title of the modal dialog. Defaults to "Add New Table".}

\item{table_types}{A character vector of table type options. Defaults to c("Fixed", "Mapping").}

\item{selected_type}{The default selected table type. Defaults to "Fixed".}
}
\value{
A Shiny modal dialog UI object.
}
\description{
This function generates a modal dialog UI object for adding a new table.
}
