% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_initialize_mapping.R
\name{initialize_map_server}
\alias{initialize_map_server}
\title{Initialize Lookup Module Server}
\usage{
initialize_map_server(id, selected_instance_url, selected_dsd_id, json_data)
}
\arguments{
\item{id}{Module ID}

\item{selected_instance_url}{A reactive expression that returns the base URL for API calls.}

\item{selected_dsd_id}{A reactive expression that returns the selected DSD ID.}

\item{json_data}{A reactive expression that returns the mapping structure.}
}
\value{
None.
}
\description{
A Shiny module server for initializing a new lookup from a selected DSD.
}
