% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_initialize_mapping.R
\name{initialize_map_ui}
\alias{initialize_map_ui}
\title{Initialize map Module UI}
\usage{
initialize_map_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
An action button to initialize a new mapping.
}
\description{
A Shiny module UI for initializing a mapping from a DSD.
}
