% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_dsd_list}
\alias{parse_dsd_list}
\title{Parse DSD List from API Response}
\usage{
parse_dsd_list(api_response)
}
\arguments{
\item{api_response}{A list representing the parsed JSON response from the API containing DSD information.}
}
\value{
A character vector of DSD IDs in the format "agencyId:id(version)".
}
\description{
Parses the API response containing Data Structure Definitions (DSDs) and extracts a list of DSD IDs in the format "agencyId:id(version)".
}
\examples{
\dontrun{
# Assuming you have the JSON response as a string
json_string <- '{"DataStructure": [...]}'
api_response <- jsonlite::fromJSON(json_string, simplifyVector = FALSE)
dsd_ids <- parse_dsd_list(api_response)
}

}
