% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_table.R
\name{parse_list_for_choices}
\alias{parse_list_for_choices}
\title{Parse a List to Generate Nested Choices for Shiny Inputs}
\usage{
parse_list_for_choices(data_list)
}
\arguments{
\item{data_list}{A list with at least two elements:
- `components`: A vector of component names.
- `representation`: A named list of representations.}
}
\value{
A named list where:
  - The `components` key contains a single element with a name and value of "components".
  - The `representation` key contains a list with the names of `representation` elements.
}
\description{
This function processes a given list to create a nested structure suitable
for the `choices` argument in Shiny `selectInput()` components. It takes
two elements from the input list: `components` and `representation`, and
organizes them into a named list for dropdown menus.
}
\examples{
input_list <- list(
  components = c("mpg", "cyl", "disp"),
  representation = list(FREQ = NULL, REF_AREA = NULL, MEASURE = NULL)
)
parse_list_for_choices(input_list)

}
