% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_row_manager.R
\name{row_manager_server}
\alias{row_manager_server}
\title{Row Manager Module Server}
\usage{
row_manager_server(
  id,
  selected_table_name,
  selected_table,
  selected_rows,
  json_data
)
}
\arguments{
\item{id}{Module ID}

\item{selected_table_name}{A reactive value containing the name of the selected table.}

\item{selected_table}{A reactive value containing the selected table data.}

\item{selected_rows}{A reactive value containing the indices of selected rows.}

\item{json_data}{A reactive value containing the JSON data.}
}
\value{
None.
}
\description{
A Shiny module server for managing rows in the selected table.
}
