% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_row_manager.R
\name{row_manager_ui}
\alias{row_manager_ui}
\title{Row Manager Module UI}
\usage{
row_manager_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI with buttons to add and delete rows.
}
\description{
A Shiny module UI for adding and deleting rows in the table.
}
