% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_dsd.R
\name{select_dsd_server}
\alias{select_dsd_server}
\title{DSD Selector Module Server}
\usage{
select_dsd_server(id, selected_instance_url)
}
\arguments{
\item{id}{Module ID}

\item{selected_instance_url}{A reactive expression that returns the base URL for API calls.}
}
\value{
A list containing a reactive expression `selected_dsd_id` that returns the selected DSD ID.
}
\description{
A Shiny module server for fetching available DSDs and handling DSD selection.
}
