% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_dsd.R
\name{select_dsd_ui}
\alias{select_dsd_ui}
\title{DSD Selector Module UI}
\usage{
select_dsd_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A `selectInput` UI element for selecting a DSD.
}
\description{
A Shiny module UI for selecting a DSD from the available DSDs.
}
