% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_fmr.R
\name{select_fmr_server}
\alias{select_fmr_server}
\title{Instance Selector Module Server}
\usage{
select_fmr_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A list containing a reactive expression `selected_instance_url` that returns the selected instance URL.
}
\description{
A Shiny module server for handling instance selection and providing the corresponding URL.
}
