% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_fmr.R
\name{select_fmr_ui}
\alias{select_fmr_ui}
\title{Instance Selector Module UI}
\usage{
select_fmr_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A `selectInput` UI element for selecting an instance.
}
\description{
A Shiny module UI for selecting the FMR instance (DEV, QA, PROD).
}
