% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_table.R
\name{select_table_server}
\alias{select_table_server}
\title{Table Selector Module Server}
\usage{
select_table_server(id, json_data, selected_instance_url)
}
\arguments{
\item{id}{Module ID}

\item{json_data}{A reactive value containing the JSON data.}

\item{selected_instance_url}{A reactive value containing FMR base URL.}
}
\value{
A list containing the reactive `selected_table_name` and an `update_table_choices` function.
}
\description{
A Shiny module server for handling table selection.
}
