% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_table.R
\name{select_table_ui}
\alias{select_table_ui}
\title{Table Selector Module UI}
\usage{
select_table_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A `selectInput` for selecting a table to edit.
}
\description{
A Shiny module UI for selecting tables from the JSON data.
}
