% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_table.R
\name{show_codelist_modal}
\alias{show_codelist_modal}
\title{Show Codelist Modal}
\usage{
show_codelist_modal(ns, codelist_df)
}
\arguments{
\item{ns}{A namespace function, typically `session$ns`, used to ensure unique IDs for Shiny inputs and outputs.}

\item{codelist_df}{A data frame representing the codelist to display. This data frame should contain
the data to be shown in the DataTable, typically with columns like `ID` and `LABEL`.}
}
\value{
None. The function is used for its side effect of displaying a modal dialog in the Shiny app.
}
\description{
Displays a modal dialog in a Shiny app with a DataTable for selecting rows from a codelist.
The modal includes a cancel button to close the dialog and a confirm button to finalize the selection.
}
\examples{
# Example usage in a Shiny app:
shinyServer(function(input, output, session) {
  ns <- session$ns
  codelist_df <- data.frame(ID = c("A1", "B2"), LABEL = c("Label A1", "Label B2"))
  show_codelist_modal(ns, codelist_df)
})

}
