% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_table.R
\name{show_define_mapping_type_modal}
\alias{show_define_mapping_type_modal}
\title{Show Define Mapping Type Modal}
\usage{
show_define_mapping_type_modal(
  ns,
  title = "Define Mapping Type",
  table_name,
  table_types = c("Fixed", "Mapping"),
  selected_type = "Fixed"
)
}
\arguments{
\item{ns}{A namespace function for unique IDs.}

\item{title}{Title of the modal dialog.}

\item{table_types}{A vector of table type options.}

\item{selected_type}{The default selected table type.}
}
\description{
This function shows a modal dialog for adding a new table, using `define_mapping_type_modal`.
}
