% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_editor.R
\name{table_editor_server}
\alias{table_editor_server}
\title{Table Editor Module Server}
\usage{
table_editor_server(id, selected_table_name, json_data)
}
\arguments{
\item{id}{Module ID}

\item{selected_table_name}{A reactive value containing the name of the selected table.}

\item{json_data}{A reactive value containing the JSON data.}
}
\value{
A list containing the reactive `selected_table` and `selected_rows`.
}
\description{
A Shiny module server for handling table editing.
}
