% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_editor.R
\name{table_editor_ui}
\alias{table_editor_ui}
\title{Table Editor Module UI}
\usage{
table_editor_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A `DTOutput` for displaying the editable table.
}
\description{
A Shiny module UI for displaying and editing the selected table.
}
