% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_manager.R
\name{table_manager_ui}
\alias{table_manager_ui}
\title{Table Manager Module UI}
\usage{
table_manager_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI with buttons to add and delete tables.
}
\description{
A Shiny module UI for adding and deleting tables in the JSON data.
}
