% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_selector.R
\name{table_selector_server}
\alias{table_selector_server}
\title{Table Selector Module Server}
\usage{
table_selector_server(id, json_data)
}
\arguments{
\item{id}{Module ID}

\item{json_data}{A reactive value containing the JSON data.}
}
\value{
A list containing the reactive `selected_table_name` and an `update_table_choices` function.
}
\description{
A Shiny module server for handling table selection.
}
