% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_selector.R
\name{table_selector_ui}
\alias{table_selector_ui}
\title{Table Selector Module UI}
\usage{
table_selector_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI output for selecting a table to edit.
}
\description{
A Shiny module UI for selecting tables from the JSON data.
}
