% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_add_table.R
\name{add_table_ui}
\alias{add_table_ui}
\title{Add Table Module UI}
\usage{
add_table_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI with a button to trigger the addition of a table.
}
\description{
A Shiny module UI for adding tables to the JSON data.
}
