% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_delete.R
\name{delete_table_ui}
\alias{delete_table_ui}
\title{Delete table Module UI}
\usage{
delete_table_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI with buttons to delete tables.
}
\description{
A Shiny module UI for deleting tables in the JSON data.
}
